feat(frontend): integrate health indicators into DevicesPage, GateDetailPage, and extract GateOverviewCard
