+++ server/frontend/src/features/gates/index.js
@@ -2,3 +2,4 @@ export { default as StatusTables } from './components/StatusTables';
 export { default as StatusTablesView } from './components/StatusTablesView';
 export { default as StatusChangedDialog } from './components/StatusChangedDialog';
 export { default as ManualStatusDialog } from './components/ManualStatusDialog';
+export { default as GateOverviewCard } from './components/GateOverviewCard';
