+++ server/frontend/src/pages/DevicesPage.jsx
@@ -2,60 +2,19 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import AppLayout from '../features/shell/components/AppLayout';
 import { useGetGatesQuery } from '../app/store/api/api';
-
-const statusInfo = (status) => {
-  switch (status) {
-    case 'OPEN': return { cls: 'status-open', label: 'Open' };
-    case 'CLOSED': return { cls: 'status-closed', label: 'Closed' };
-    default: return { cls: 'status-oos', label: 'OOS' };
-  }
-};
-
-const priorityClass = (level) => {
-  switch (level) {
-    case 0: return 'priority-low';
-    case 1: return 'priority-medium';
-    case 2: return 'priority-high';
-    case 3: return 'priority-critical';
-    default: return 'priority-low';
-  }
-};
-
-const confidenceColor = (conf) => {
-  if (conf == null) return 'var(--text-secondary)';
-  if (conf >= 90) return 'var(--green-600)';
-  if (conf >= 70) return 'var(--amber-600)';
-  return 'var(--red-600)';
-};
-
-const stateConfirmationInfo = (sc) => {
-  switch (sc) {
-    case 'WORKER_CONFIRMED_SINGLE': return { cls: 'status-open', label: '1 worker' };
-    case 'WORKER_CONFIRMED_MULTI': return { cls: 'status-closed', label: '2+ workers' };
-    case 'WORKER_CONFIRMED_ALL': return { cls: 'status-closed', label: 'All workers' };
-    case 'WORKER_CONFLICT': return { cls: 'status-oos', label: 'Conflict' };
-    case 'UNCONFIRMED': return { cls: 'status-none', label: 'Unconfirmed' };
-    default: return { cls: 'status-none', label: sc || '—' };
-  }
-};
-
-function getTimeAgo(timestamp) {
-  const date = new Date(timestamp);
-  const now = new Date();
-  const secondsAgo = Math.floor((now - date) / 1000);
-  if (secondsAgo < 60) return `${secondsAgo}s ago`;
-  const minutes = Math.floor(secondsAgo / 60);
-  if (minutes < 60) return `${minutes}m ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  if (days === 1) return 'yesterday';
-  return `${days}d ago`;
-}
+import { useAppSelector } from '../app/store';
+import GateOverviewCard from '../features/gates/components/GateOverviewCard';
+import { HealthBadge } from '../features/health';
 
 const DevicesPage = () => {
   const navigate = useNavigate();
   const { data: gates = [], isLoading, error } = useGetGatesQuery();
+  const healthBySenseGateId = useAppSelector((state) => state.health.bySenseGateId);
+
+  const gateIds = new Set(gates.map((g) => g.id));
+  const unmappedEntries = Object.entries(healthBySenseGateId).filter(
+    ([senseGateId]) => !gateIds.has(Number(senseGateId))
+  );
 
   if (isLoading) {
     return (
@@ -95,95 +54,42 @@ const DevicesPage = () => {
         </p>
       </div>
       <div className="gate-overview-grid">
-        {gates.map((gate) => {
-          const si = statusInfo(gate.status);
-          const pc = priorityClass(gate.priority);
-          const cc = confidenceColor(gate.confidence);
-          const stateConf = stateConfirmationInfo(gate.stateConfirmation);
-          return (
-            <div
-              key={gate.id}
-              className="gate-overview-card"
-              onClick={() => navigate(`/gates/${gate.id}`)}
-            >
-              <div className="gate-overview-card-header">
-                <div style={{ display: 'flex', alignItems: 'center', gap: '8px', minWidth: 0 }}>
-                  <span className="gate-id">#{gate.id}</span>
-                  <span
-                    style={{
-                      fontSize: '14px',
-                      fontWeight: 600,
-                      color: 'var(--text)',
-                      overflow: 'hidden',
-                      textOverflow: 'ellipsis',
-                      whiteSpace: 'nowrap',
-                    }}
-                  >
-                    {gate.location || 'Unnamed'}
-                  </span>
-                </div>
-                <div style={{ display: 'flex', alignItems: 'center', gap: '6px', flexShrink: 0 }}>
-                  {gate.manualOverride && <span className="manual-pill">Manual</span>}
-                  <span className={`status-badge ${si.cls}`}>
-                    <span className="status-dot" />
-                    {si.label}
-                  </span>
-                </div>
-              </div>
-              <div className="gate-overview-card-body">
-                <div>
-                  <span className="gate-overview-info-label">Confidence</span>
-                  <span className="gate-overview-info-value" style={{ display: 'inline-flex', alignItems: 'center', gap: '6px' }}>
-                    <span
-                      style={{
-                        width: '8px',
-                        height: '8px',
-                        borderRadius: '50%',
-                        background: cc,
-                        display: 'inline-block',
-                        flexShrink: 0,
-                      }}
-                    />
-                    {gate.confidence != null ? `${gate.confidence}%` : '—'}
-                  </span>
-                </div>
-                <div>
-                  <span className="gate-overview-info-label">Height above NN</span>
-                  <span className="gate-overview-info-value">
-                    {gate.heightAboveNN != null ? `${gate.heightAboveNN} m` : '—'}
-                  </span>
-                </div>
-                <div className={pc}>
-                  <span className="gate-overview-info-label">Priority</span>
-                  <span className="gate-overview-info-value" style={{ display: 'inline-flex', alignItems: 'center' }}>
-                    <span className="priority-dot" />
-                    {gate.priority != null ? gate.priority : '—'}
-                  </span>
-                </div>
-                <div>
-                  <span className="gate-overview-info-label">Last Update</span>
-                  <span className="gate-overview-info-value last-update">
-                    {gate.lastTimeStamp ? getTimeAgo(gate.lastTimeStamp) : '—'}
-                  </span>
-                </div>
-                <div>
-                  <span className="gate-overview-info-label">State Confirmation</span>
-                  <span className={`status-badge ${stateConf.cls}`} style={{ marginTop: '2px' }}>
-                    <span className="status-dot" />
-                    {stateConf.label}
-                  </span>
-                </div>
-              </div>
-              <div className="gate-overview-card-footer">
-                <span style={{ color: 'var(--text-secondary)' }}>
-                  Device: <span className="gate-id">{gate.deviceId ?? '—'}</span>
-                </span>
-                <span className="action-link">View Details →</span>
-              </div>
-            </div>
-          );
-        })}
+        {gates.map((gate) => (
+          <GateOverviewCard
+            key={gate.id}
+            gate={gate}
+            onClick={() => navigate(`/gates/${gate.id}`)}
+          />
+        ))}
       </div>
+      {unmappedEntries.length > 0 && (
+        <div style={{ marginTop: '24px' }}>
+          <h2 style={{ fontSize: '20px', fontWeight: 700, margin: 0 }}>
+            Unmapped Health Devices
+          </h2>
+          <p style={{ fontSize: '14px', color: 'var(--text-secondary)', marginTop: '4px' }}>
+            Health data received for devices not in the gate list
+          </p>
+          <div className="gate-overview-grid">
+            {unmappedEntries.map(([senseGateId, health]) => (
+              <div key={senseGateId} className="gate-overview-card" style={{ opacity: 0.7 }}>
+                <div className="gate-overview-card-header">
+                  <span className="gate-id">#{senseGateId}</span>
+                  <span style={{ fontSize: '14px', fontWeight: 600, color: 'var(--text-secondary)' }}>
+                    Unmapped Device
+                  </span>
+                </div>
+                <div className="gate-overview-card-body">
+                  <div>
+                    <span className="gate-overview-info-label">Health</span>
+                    <HealthBadge health={health} />
+                  </div>
+                </div>
+              </div>
+            ))}
+          </div>
+        </div>
+      )}
     </AppLayout>
   );
 };
